(fix) Fixes unwind segue from ARScan main menu.
VC still not instantiating on the proper Nav stack from scan completion to AddProductShiftableVC
