+++ ScannAR/ScannAR/ARScanner Scene/ViewControllers/ARScanViewController.swift
@@ -5,7 +5,7 @@
 //  Created by Joshua Kaunert on 3/20/19.
 //  Copyright © 2019 ScannAR Team. All rights reserved.
 //
-
+import Foundation
 import UIKit
 import SceneKit
 import ARKit
@@ -337,21 +337,31 @@ class ARScanViewController: UIViewController, ARSCNViewDelegate, ARSessionDelega
                         }
                        
                         // MARK: - Do not remove commented out code
-//                        let viewController = UIStoryboard(name: "ScannARMainViewController", bundle: nil).instantiateViewController(withIdentifier: "AddProductViewControllerSB") as! AddProductViewController
                         
+                        let viewController = UIStoryboard(name: "ScannARMainViewController", bundle: nil).instantiateViewController(withIdentifier: "AddProductViewContainerVC") as! AddProductContainerViewController
+                        let navController = UINavigationController(rootViewController: viewController)
                         let rotatedScreenshot = self.objectScreenshot?.imageRotatedByDegrees(degrees: 90, flip: false)
                         self.previewImage = rotatedScreenshot
                         
-                        //viewController.bestBoxSize = self.boundingBoxSize
-                        //print(self.boundingBoxSize)
+                        viewController.bestBoxSize = self.boundingBoxSize
+                        viewController.previewImage = self.previewImage!
+//                        viewController.height = self.boundingBoxSize.height ?? 0.0
+//                        viewController.width = self.boundingBoxSize.width ?? 0.0
+//                        viewController.length = self.boundingBoxSize.length ?? 0.0
+                        viewController.name = self.scannedObjectName
                         
-//                        let transition: CATransition = CATransition()
-//                        transition.duration = 0.7
-//                        transition.timingFunction = CAMediaTimingFunction(name: CAMediaTimingFunctionName.easeInEaseOut)
-//                        transition.type = CATransitionType.fade
+                        print(self.boundingBoxSize)
+                        
+                        let transition: CATransition = CATransition()
+                        transition.duration = 0.7
+                        transition.timingFunction = CAMediaTimingFunction(name: CAMediaTimingFunctionName.easeInEaseOut)
+                        transition.type = CATransitionType.fade
 //
-//                        self.navigationController!.view.layer.add(transition, forKey: nil)
-                        self.performSegue(withIdentifier: "unwindSegueToAddProductVC", sender: self)
+                        self.navigationController!.view.layer.add(transition, forKey: nil)
+                        self.present(navController, animated:true, completion: nil)
+                        //self.present(viewController, animated: false)
+//                        self.performSegue(withIdentifier: "unwindToAddProductVC", sender: self)
+                        
 
                 }))
